patrimo/requirements.txt
fastapi==0.111.0
uvicorn==0.29.0
openpyxl==3.1.2
httpx==0.27.0
apscheduler==3.10.4
python-dotenv==1.0.1
